Rename misc to variables
