+++ src/components/variables/VariableTitle/VariableTitle.js
